Pickup Widget Progress
Forgot to commit.  Found a pack with some cool textures for the UI and implemented a pickup widget.  I also found some textures randomly in some other packs that I was able to grab that were free.  Added Sci Fi fonts as well.  Logic for showing and hiding the pickup widget is complete with a failsafe.